+++ pytest_factoryboy/__init__.py
@@ -1,7 +1,8 @@
 """pytest-factoryboy public API."""
 from .fixture import LazyFixture, register
 
-__version__ = "2.3.0"
+# TODO: Remove this, it's managed by poetry
+__version__ = "2.4.0"
 
 
 __all__ = ("register", "LazyFixture")