+++ pytest_factoryboy/fixture.py
@@ -1,12 +1,25 @@
 """Factory boy fixture integration."""
 from __future__ import annotations
 
+import functools
 import inspect
 import os
 import sys
 from dataclasses import dataclass
 from inspect import signature
-from typing import TYPE_CHECKING, overload
+from typing import (
+    TYPE_CHECKING,
+    Any,
+    Callable,
+    Collection,
+    Generic,
+    Iterable,
+    Mapping,
+    Type,
+    TypeVar,
+    cast,
+    overload,
+)
 
 import factory
 import factory.builder
@@ -14,9 +27,11 @@ import factory.declarations
 import factory.enums
 import inflection
 from factory.declarations import NotProvided
+from typing_extensions import ParamSpec, TypeAlias
 
-from .codegen import FixtureDef, make_fixture_model_module, upgrade_module
+from .codegen import upgrade_module
 from .compat import PostGenerationContext
+from .fixturegen import create_fixture
 
 if TYPE_CHECKING:
     from types import FrameType, ModuleType
@@ -26,9 +41,12 @@ if TYPE_CHECKING:
     from factory.builder import BuildStep
     from factory.declarations import PostGeneration, PostGenerationContext
 
-    FactoryType = type[factory.Factory]
-    T = TypeVar("T")
-    F = TypeVar("F", bound=FactoryType)
+    from .plugin import Request as FactoryboyRequest
+
+FactoryType: TypeAlias = Type[factory.Factory]
+F = TypeVar("F", bound=FactoryType)
+T = TypeVar("T")
+P = ParamSpec("P")
 
 
 def should_rewrite_source() -> bool:
@@ -211,92 +229,123 @@ def _register_new(
     assert not factory_class._meta.abstract, "Can't register abstract factories."
     assert factory_class._meta.model is not None, "Factory model class is not specified."
 
-    fixture_defs: list[FixtureDef] = []
+    model_name = get_model_name(factory_class) if _name is None else _name
 
-    model_name = get_model_name(factory_class) if name is None else name
-    factory_name = get_factory_name(factory_class)
-
-    deps = get_deps(factory_class, model_name=model_name)
-    related: list[str] = []
-
-    for attr, value in factory_class._meta.declarations.items():
-        args = []
-        attr_name = SEPARATOR.join((model_name, attr))
-
-        if isinstance(value, (factory.SubFactory, factory.RelatedFactory)):
-            value = factory_kwargs.get(attr, value)
-            subfactory_class = value.get_factory()
-            subfactory_deps = get_deps(subfactory_class, factory_class)
-
-            args = list(subfactory_deps)
-            if isinstance(value, factory.RelatedFactory):
-                related_model = get_model_name(subfactory_class)
-                args.append(related_model)
-                related.append(related_model)
-                related.append(attr_name)
-                related.extend(subfactory_deps)
-
-            if isinstance(value, factory.SubFactory):
-                args.append(inflection.underscore(subfactory_class._meta.model.__name__))
-
-            fixture_defs.append(
-                FixtureDef(
-                    name=attr_name,
-                    function_name="subfactory_fixture",
-                    function_kwargs={"factory_class": subfactory_class},
-                    deps=args,
-                )
-            )
-            continue
-
-        if isinstance(value, factory.PostGeneration):
-            default_value = None
-        elif isinstance(value, factory.PostGenerationMethodCall):
-            default_value = value.method_arg
-        else:
-            default_value = value
-
-        value = factory_kwargs.get(attr, default_value)
-
-        if isinstance(value, LazyFixture):
-            args = value.args
-
-        fixture_defs.append(
-            FixtureDef(
-                name=attr_name,
-                function_name="attr_fixture",
-                function_kwargs={"value": value},
-                deps=args,
-            )
-        )
-
-    if factory_name not in _caller_locals:
-        fixture_defs.append(
-            FixtureDef(
-                name=factory_name,
-                function_name="factory_fixture",
-                function_kwargs={"factory_class": factory_class},
-            )
-        )
-
-    fixture_defs.append(
-        FixtureDef(
-            name=model_name,
-            function_name="model_fixture",
-            function_kwargs={"factory_name": factory_name},
-            deps=deps,
-            related=related,
+    fixture_defs = dict(
+        generate_fixtures(
+            factory_class=factory_class, model_name=model_name, overrides=kwargs, caller_locals=_caller_locals
         )
     )
-
-    generated_module = make_fixture_model_module(model_name, fixture_defs)
-
-    for fixture_def in fixture_defs:
-        exported_name = fixture_def.name
-        fixture_function = getattr(generated_module, exported_name)
-        inject_into_caller(exported_name, fixture_function, _caller_locals)
+    for name, fixture in fixture_defs.items():
+        inject_into_caller(name, fixture, _caller_locals)
 
     return factory_class
+
+
+def generate_fixtures(
+    factory_class: FactoryType, model_name: str, overrides: Mapping[str, Any], caller_locals: Mapping[str, Any]
+) -> Iterable[tuple[str, Callable[..., Any]]]:
+    """Generate all the FixtureDefs for the given factory class."""
+    factory_name = get_factory_name(factory_class)
+
+    related: list[str] = []
+    for attr, value in factory_class._meta.declarations.items():
+        value = overrides.get(attr, value)
+        attr_name = SEPARATOR.join((model_name, attr))
+        yield (
+            attr_name,
+            make_declaration_fixturedef(
+                attr_name=attr_name,
+                value=value,
+                factory_class=factory_class,
+                related=related,
+            ),
+        )
+
+    if factory_name not in caller_locals:
+        yield (
+            factory_name,
+            create_fixture_with_related(
+                name=factory_name,
+                function=functools.partial(factory_fixture, factory_class=factory_class),
+            ),
+        )
+
+    deps = get_deps(factory_class, model_name=model_name)
+    yield (
+        model_name,
+        create_fixture_with_related(
+            name=model_name,
+            function=functools.partial(model_fixture, factory_name=factory_name),
+            fixtures=deps,
+            related=related,
+        ),
+    )
+
+
+def create_fixture_with_related(
+    name: str,
+    function: Callable[P, T],
+    fixtures: Collection[str] | None = None,
+    related: Collection[str] | None = None,
+) -> Callable[P, T]:
+    if related is None:
+        related = []
+    f = create_fixture(name=name, function=function, fixtures=fixtures)
+
+    # We have to set the `_factoryboy_related` attribute to the original function, since
+    # FixtureDef.func will provide that one later when we discover the related fixtures.
+    f.__pytest_wrapped__.obj._factoryboy_related = related  # type: ignore[attr-defined]
+    return f
+
+
+def make_declaration_fixturedef(
+    attr_name: str,
+    value: Any,
+    factory_class: FactoryType,
+    related: list[str],
+) -> Callable[..., Any]:
+    """Create the FixtureDef for a factory declaration."""
+    if isinstance(value, (factory.SubFactory, factory.RelatedFactory)):
+        subfactory_class = value.get_factory()
+        subfactory_deps = get_deps(subfactory_class, factory_class)
+
+        args = list(subfactory_deps)
+        if isinstance(value, factory.RelatedFactory):
+            related_model = get_model_name(subfactory_class)
+            args.append(related_model)
+            related.append(related_model)
+            related.append(attr_name)
+            related.extend(subfactory_deps)
+
+        if isinstance(value, factory.SubFactory):
+            args.append(inflection.underscore(subfactory_class._meta.model.__name__))
+
+        return create_fixture_with_related(
+            name=attr_name,
+            function=functools.partial(subfactory_fixture, factory_class=subfactory_class),
+            fixtures=args,
+        )
+
+    deps: list[str]  # makes mypy happy
+    if isinstance(value, factory.PostGeneration):
+        value = None
+        deps = []
+    elif isinstance(value, factory.PostGenerationMethodCall):
+        value = value.method_arg
+        deps = []
+    elif isinstance(value, LazyFixture):
+        value = value
+        deps = value.args
+    else:
+        value = value
+        deps = []
+
+    return create_fixture_with_related(
+        name=attr_name,
+        function=functools.partial(attr_fixture, value=value),
+        fixtures=deps,
+    )
 
 
 new_register_signature = signature(_register_new)
@@ -323,11 +372,18 @@ def inject_into_caller(name: str, function: Callable[..., Any], locals_: dict[st
 
 def get_model_name(factory_class: FactoryType) -> str:
     """Get model fixture name by factory."""
-    return (
-        inflection.underscore(factory_class._meta.model.__name__)
-        if not isinstance(factory_class._meta.model, str)
-        else factory_class._meta.model
-    )
+    # "class AuthorFactory(...):" -> "author"
+    suffix = "Factory"
+    if factory_class.__name__.endswith(suffix):
+        factory_name = factory_class.__name__[: -len(suffix)]
+        return inflection.underscore(factory_name)
+    else:
+        raise ValueError(
+            f"Factory {factory_class} does not follow the '<model_name>Factory' naming convention and "
+            "pytest-factoryboy cannot automatically determine the fixture name.\n"
+            f"Please use the naming convention '<model_name>Factory' or explicitly set the fixture name "
+            "using `@register(_name=...)`."
+        )
 
 
 def get_factory_name(factory_class: FactoryType) -> str:
@@ -363,7 +419,7 @@ def get_deps(
     ]
 
 
-def evaluate(request: SubRequest, value: LazyFixture | Any) -> Any:
+def evaluate(request: SubRequest, value: LazyFixture[T] | T) -> T:
     """Evaluate the declaration (lazy fixtures, etc)."""
     return value.evaluate(request) if isinstance(value, LazyFixture) else value
 
@@ -371,7 +427,7 @@ def evaluate(request: SubRequest, value: LazyFixture | Any) -> Any:
 def model_fixture(request: SubRequest, factory_name: str) -> Any:
     """Model fixture implementation."""
     raise_if_upgrade_necessary()
-    factoryboy_request = request.getfixturevalue("factoryboy_request")
+    factoryboy_request: FactoryboyRequest = request.getfixturevalue("factoryboy_request")
 
     # Try to evaluate as much post-generation dependencies as possible
     factoryboy_request.evaluate(request)
@@ -379,13 +435,15 @@ def model_fixture(request: SubRequest, factory_name: str) -> Any:
     assert request.fixturename  # NOTE: satisfy mypy
     fixture_name = request.fixturename
     prefix = "".join((fixture_name, SEPARATOR))
-    # NOTE: following type hinting is required, because of `mypy` bug.
-    # Reference: https://github.com/python/mypy/issues/2477
-    factory_class: factory.base.FactoryMetaClass = request.getfixturevalue(factory_name)
+
+    factory_class: FactoryType = request.getfixturevalue(factory_name)
 
     # Create model fixture instance
-    class Factory(factory_class):
-        pass
+    Factory: FactoryType = cast(FactoryType, type("Factory", (factory_class,), {}))
+    # equivalent to:
+    # class Factory(factory_class):
+    #     pass
+    # it just makes mypy understand it.
 
     Factory._meta.base_declarations = {
         k: v
@@ -533,10 +591,10 @@ def get_caller_locals(depth: int = 2) -> dict[str, Any]:
     return get_caller_frame(depth + 1).f_locals
 
 
-class LazyFixture:
+class LazyFixture(Generic[T]):
     """Lazy fixture."""
 
-    def __init__(self, fixture: FixtureFunction | str) -> None:
+    def __init__(self, fixture: Callable[..., T] | str) -> None:
         """Lazy pytest fixture wrapper.
 
         :param fixture: Fixture name or callable with dependencies.
@@ -548,7 +606,7 @@ class LazyFixture:
         else:
             self.args = [self.fixture]
 
-    def evaluate(self, request: SubRequest) -> Any:
+    def evaluate(self, request: SubRequest) -> T:
         """Evaluate the lazy fixture.
 
         :param request: pytest request object.
@@ -558,4 +616,4 @@ class LazyFixture:
             kwargs = {arg: request.getfixturevalue(arg) for arg in self.args}
             return self.fixture(**kwargs)
         else:
-            return request.getfixturevalue(self.fixture)
+            return cast(T, request.getfixturevalue(self.fixture))